Renamed serverProtocol to clientProtocol Add more commands to the clientProtocol Renamed add_item to sell_item
The auction now supports real time bidding and if the deadline passes, you cannot bid on the item anymore. Added SellingThread to support this.

Add the SanityCheck to see that every bidder always gets the item of his best utility.

Server now has its own serverProtocol and can listen to our commands.

Made a minor change to Console.java as well.
Fixed some bug with Auction.java where you are told to be matched to the Dummy if you bid 3 times (increasing the price with each time) on the same item

Group bid and change bid are not implemented yet.